+++ src/main/java/kitchenpos/order/domain/OrderLineItem.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
+
+import kitchenpos.order.domain.Order;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
