+++ src/main/java/kitchenpos/order/presentation/OrderRestController.java
@@ -1,9 +1,9 @@
-package kitchenpos.ui;
+package kitchenpos.order.presentation;
 
-import kitchenpos.application.OrderService;
-import kitchenpos.domain.Order;
-import kitchenpos.ui.dto.OrderCreateRequest;
-import kitchenpos.ui.dto.OrderStatusChangeRequest;
+import kitchenpos.order.application.OrderService;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.presentation.dto.OrderCreateRequest;
+import kitchenpos.order.presentation.dto.OrderStatusChangeRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
