+++ src/main/java/kitchenpos/order/presentation/dto/OrderStatusChangeRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.ui.dto;
+package kitchenpos.order.presentation.dto;
 
 public class OrderStatusChangeRequest {
