+++ src/test/java/kitchenpos/order/domain/repository/OrderRepositoryTest.java
@@ -1,6 +1,7 @@
-package kitchenpos.domain.repository;
+package kitchenpos.order.domain.repository;
 
-import kitchenpos.domain.Order;
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.repository.OrderRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
